feat(sourceprocessor)!: allow templating of source host field
This change replaces `source_host_key` configuration option
with `source_host` option.
This is for consistency, to allow templating of
source host field in the same way that source name
and source category fields can be templated.

BREAKING CHANGE: `source_host_key` option is removed in favor of `source_host` option.
To upgrade, take the value of `source_host_key`, sorround it with `'%{` and `}'`,
and put it in `source_host` option.

For example: `source_host_key: k8s.pod.hostname`
should be changed to `source_host: '%{k8s.pod.hostname}'`