docs(configuration): avoid having to explain `metadata_attributes`
This rewrites the docs on configuring source host
to avoid having to explain the quirky `metadata_attributes`
property on the Sumo Logic exporter.
Also, the method of setting the `_sourceHost`
attribute directly is in line with the docs for source category.
Also, the user can easily see the connection between
setting a `_sourceHost` attribute and using it in Sumo queries.