feat(k8sprocessor)!: remove `extract.tags` configuration
This change removes the configuration property `extract.tags`.
This way, the user is not able to customize the names
of the attributes for the extracted metadata.
This can be done using a resourceprocessor later in the pipeline.
To update the configuration, remove the `extract.tags`
property and rename the elements of `extract.metadata` property,
changing e.g. `containerName` to `k8s.container.name`.
The old configuration elements in `extract.metadata`
like `containerName` etc. are now deprecated and will be supported
for a couple releases before they are removed.