feat(sumologicexporter): do not send empty OTLP requests (#660)
Currently, when en empty metrics or traces dataset is passed to the Sumo Logic exporter, an empty-bodied request is sent to the backend.
This behavior is in fact in line with what the upstream OTLP HTTP exporter does for all signals - logs, metrics and traces.

This change modifies this behavior for metrics and traces so that the empty request is not sent. This is to prevent an unnecessary network call.

For logs, the current behavior is to not send the empty request when an empty dataset is passed to the exporter, and there is no change to this behavior.