docs: simplify Docker run instructions
Fixes #790

We probably don't need to show the output
of the `docker run` command at all;
we don't show it in other examples for running the binary either.
I have also changed the target path of the config file for simplicity
and split the command into multiline lines
so that it is visible in the viewport without scrolling.