fix(sumologicexporter): fix links to upgrade docs
We were trying to be helpful by helping the user
migrate off unsupported properties,
but links that point to nowhere do not help much.
This changes the links to point straight to the upgrade docs
instead of release notes.
I also point to the `main` branch of the upgrade docs,
because there have been, and there might be more,
changes/improvements to these docs.