feat(receiver/sqlquery): go back to using upstream receiver
We switched to using a fork of the SQL Query receiver in v0.78.0
to introduce logs support before it was available in upstream.
Now that upstream has logs support, we can remove the override.

This change is actually a bit belated. We could and should've done it
in v0.80.0, as the upstream already included logs support in it.