Add --opamp-config flag
The --opamp-config flag is used to specify an OpAmp configuration file.
The file will be updated according to the OpAmp server's remote
management directives.

Since this method of configuration is so different from the
configuration expressed by the --config flag, --opamp-config and
--config cannot be combined.

This change marks the second time it was necessary to patch main.go, and
so I've created an adhoc patch queue by prefixing patches with numbers
that indicate the order the patches should be applied in.

Signed-off-by: Eric Chlebek <eric@sensu.io>