Add OpAMP config provider
This commit adds a configuration watcher that's intended to be used in
concert with the --opamp-config flag. It's implemented with fsnotify,
with a custom polling fallback implemented when the underlying
filesystem doesn't support notification, like with NFS.

The polling watcher uses mtime as a signal that a file has changed.

Signed-off-by: Eric Chlebek <eric@sensu.io>