feat(opamp): Add OpAmp configuration provider (#1287)
feat(opamp): Add OpAmp configuration provider

The OpAmp configuration provider reads a yaml configuration file,
extracting a top-level field named `remote_configuration_directory`. The
remote configuration directory is expected to contain one or more files
with an extension of `yaml`. After the directory is retrieved, the glob
provider is used to read the yaml files from it.

The OpAmp configuration provider will only activate when a user supplies
the --opamp-config flag, but is not hooked up yet, as the opampprovider
module must first be published.

Signed-off-by: Eric Chlebek <eric@sensu.io>