Upgrade opamp-go to tip
Upgrade to an untagged opamp-go in order to take advantage of the recent
support for HTTP redirection that has been added.

The project has broken several things since 0.10.x, and this breakage
has been reflected as minor changes to the opamp extension as well.

opamp-go should be upgraded to a tagged version on a subsequent release.

Signed-off-by: Eric Chlebek <eric@sensu.io>