ci: fix packaging trigger for releases
Our reusable workflow would unconditionally pass the workflow id as the
parameter to the triggered packaging workflow.
This causes the packaging workflow to fetch artifacts from the
workflow with that id, which is correct for dev builds, but incorrect
for release builds. Fix this by making the workflow id an input.