fix(scripts): restore install scripts temporarily
A recent change deleted the install scripts from this repository. The
scripts are still attached to releases, which is how they're supposed to
be consumed by users. However, it turns out that some script links in
our documentation incorrectly refer to them by their content in `main`.

Restore the scripts until those urls are corrected.