Add --delete-tag to otelcol-config
This commit adds the delete-tag action to otelcol-config.

The delete-tag action deletes a key out of the collector_fields object
in conf.d settings. If the --override flag is provided, then the action
deletes the tag out of conf.d overrides.

If a user has manually added a tag to conf.d, and tries to delete it
with this tool in override mode, then an error will be returned, and
otelcol-config will refuse to proceed.

Signed-off-by: Eric Chlebek <echlebek@sumologic.com>