Don't use race detector for jobreceiver tests
This commit overrides the race detection in the jobreceiver tests. While
this prevents us from catching data races in the jobreceiver in CI, it
also keeps CI from failing sporadically on Darwin.

There's something amiss in the jobreceiver tests and it causes CI to
fail quite regularly, almost always on Darwin. I suspect there is some
concurrency bug in the test harness that is more likely to deadlock when
the race detector is enabled. (The race detector slows everything down,
and that can make deadlocks more likely to happen by exposing timing
bugs).

Signed-off-by: Eric Chlebek <echlebek@sumologic.com>