chore(ci): pkg trigger & release workflow updates
Packaging is now only triggered from Dev & PR builds. The only input
sent when triggering packaging is now the Workflow Run ID of the
workflow run that packaging should pull its artifacts from.

The version of otelcol-sumo & otelcol-config is now determined using a
combination of the otelcol_version in
`otelcolbuilder/.otelcol-builder.yaml` and git tags.

Releases are no longer triggered by the creation of a tag. A release can
now be triggered via the new 'Publish release' workflow. The new release
workflow requires the ID of a workflow run that contains the artifacts
to be released.

The Git SHA used to build the otelcol-sumo binary is now included in the
version of the --version flag.

Signed-off-by: Justin Kolberg <amd.prophet@gmail.com>