Use staticcheck instead of golangci-lint (#1688)
This commit changes the definition of the "golint" make target. It now
uses staticcheck instead of golangci-lint.

Most of the value in golangci-lint is derived from staticcheck, and
staticcheck on its own runs in about 25% of the time as golangci-lint.
So, there is little reason to use golangci-lint.

Additionally, golangci-lint configures staticcheck in a non-standard
way. As such, some lint errors were masked. They have been fixed in this
commit as well.

Signed-off-by: Eric Chlebek <echlebek@sumologic.com>